(refactor) moved pre-check-in auth to shared location
