+++ src/applications/check-in/actions/authentication/index.js
@@ -1,0 +1,11 @@
+export const SET_SESSION = 'SET_SESSION';
+
+export const createSetSession = ({ token, permissions }) => {
+  return {
+    type: SET_SESSION,
+    payload: {
+      token,
+      permissions,
+    },
+  };
+};
