+++ src/applications/check-in/actions/pre-check-in/pre-check-in.actions.unit.spec.js
@@ -1,8 +1,6 @@
 import { expect } from 'chai';
 
 import {
-  SET_SESSION,
-  createSetSession,
   recordAnswer,
   RECORD_ANSWER,
   setVeteranData,
@@ -11,20 +9,6 @@ import {
 
 describe('pre-check-in', () => {
   describe('actions', () => {
-    describe('createSetSession', () => {
-      it('should return correct action', () => {
-        const action = createSetSession({});
-        expect(action.type).to.equal(SET_SESSION);
-      });
-      it('should return correct structure', () => {
-        const action = createSetSession({
-          token: 'some-token',
-          permissions: 'some-permission',
-        });
-        expect(action.payload.token).to.equal('some-token');
-        expect(action.payload.permissions).to.equal('some-permission');
-      });
-    });
     describe('recordAnswer', () => {
       it('should return correct action', () => {
         const action = recordAnswer({});
