+++ src/applications/check-in/reducers/authentication/index.js
@@ -1,0 +1,12 @@
+const setSessionHandler = (state, action) => {
+  return {
+    ...state,
+    context: {
+      ...state.context,
+      token: action.payload.token,
+      permissions: action.payload.permissions,
+    },
+  };
+};
+
+export { setSessionHandler };
