+++ src/applications/check-in/reducers/pre-check-in/pre-check-in.reducers.unit.spec.js
@@ -1,15 +1,10 @@
 import { expect } from 'chai';
 
-import {
-  recordAnswerHandler,
-  setSessionHandler,
-  setVeteranDataHandler,
-} from './index';
+import { recordAnswerHandler, setVeteranDataHandler } from './index';
 
 import { updateFormHandler } from '../navigation';
 
 import {
-  createSetSession,
   recordAnswer,
   setVeteranData,
   updateFormAction,
@@ -21,40 +16,6 @@ import appReducer from '../index';
 
 describe('check in', () => {
   describe('pre-check-in reducers', () => {
-    describe('createSetSession', () => {
-      describe('setSessionHandler', () => {
-        it('should return form structure', () => {
-          const action = createSetSession({
-            token: 'some-token',
-            permissions: 'some-permission',
-          });
-          const state = setSessionHandler({}, action);
-          expect(state.context).haveOwnProperty('token');
-          expect(state.context).haveOwnProperty('permissions');
-        });
-
-        it('should set session context data', () => {
-          const action = createSetSession({
-            token: 'some-token',
-            permissions: 'some-permission',
-          });
-          const state = setSessionHandler({}, action);
-          expect(state.context.token).to.equal('some-token');
-          expect(state.context.permissions).to.equal('some-permission');
-        });
-      });
-      describe('reducer is called; finds the correct handler', () => {
-        it('should set correct data', () => {
-          const action = createSetSession({
-            token: 'some-token',
-            permissions: 'some-permission',
-          });
-          const state = appReducer.checkInData(undefined, action);
-          expect(state.context.token).to.equal('some-token');
-          expect(state.context.permissions).to.equal('some-permission');
-        });
-      });
-    });
     describe('recordAnswer', () => {
       describe('recordAnswerHandler', () => {
         it('should return form structure', () => {
